extract shared default values to a core module to minimize the dependencies of natchez-http4s-mtl
