+++ .github/workflows/ci.yml
@@ -83,11 +83,11 @@ jobs:
 
       - name: Make target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
-        run: mkdir -p modules/mtl/.native/target modules/http4s/.jvm/target modules/http4s/.js/target modules/http4s/.native/target modules/mtl/.js/target modules/mtl/.jvm/target project/target
+        run: mkdir -p modules/mtl/.native/target modules/http4s/.jvm/target modules/http4s/.js/target modules/core/.native/target modules/core/.js/target modules/core/.jvm/target modules/http4s/.native/target modules/mtl/.js/target modules/mtl/.jvm/target project/target
 
       - name: Compress target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
-        run: tar cf targets.tar modules/mtl/.native/target modules/http4s/.jvm/target modules/http4s/.js/target modules/http4s/.native/target modules/mtl/.js/target modules/mtl/.jvm/target project/target
+        run: tar cf targets.tar modules/mtl/.native/target modules/http4s/.jvm/target modules/http4s/.js/target modules/core/.native/target modules/core/.js/target modules/core/.jvm/target modules/http4s/.native/target modules/mtl/.js/target modules/mtl/.jvm/target project/target
 
       - name: Upload target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
